+++ src/sys/io/mod.rs
@@ -15,19 +15,21 @@
 #![cfg_attr(feature = "safe_io", forbid(unsafe_code))]
 
 mod duplex; // IoDuplex
-
 #[cfg(any(feature = "std", all(not(feature = "std"), feature = "io")))]
 #[cfg_attr(nightly_doc, doc(cfg(any(feature = "std", all(not(feature = "std"), feature = "io")))))]
 mod io; // Io
-
-#[cfg(not(feature = "std"))]
-mod no_std_define;
-#[cfg(feature = "std")]
-mod std_reexports;
+cfg_select! {
+    feature = "std" => { mod std_reexports; }
+                  _ => { mod no_std_define; }
+}
+mod text; // TextOut // WIP
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::duplex::*;
+        pub use super::{
+            duplex::*,
+            text::*,
+        };
 
         #[cfg(any(feature = "std", all(not(feature = "std"), feature = "io")))]
         pub use super::io::*;